Use long variables to store total byte counts instead of int variables, the latter will overflow with files larger than 2GB

+++ src/org/jets3t/service/impl/rest/httpclient/RepeatableRequestEntity.java
@@ -49,7 +49,7 @@ public class RepeatableRequestEntity implements RequestEntity {
     private String contentType = null;
     private long contentLength = 0;
     
-    int bytesWritten = 0;    
+    long bytesWritten = 0;    
     private IRepeatableInputStream repeatableInputStream = null;
     private ProgressMonitoredInputStream progressMonitoredIS = null;
